// frontend/src/api/generated/model/aiotBackendApiDeviceGroupV1DeviceGroup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nunu Example API
 * This is a sample server celler server.
 * OpenAPI spec version: 1.0.0
 */

export interface AiotBackendApiDeviceGroupV1DeviceGroup {
  createdAt?: string
  deletedAt?: string
  description?: string
  groupUuid?: string
  name?: string
  rule?: string
  type?: string
  updatedAt?: string
}